move tip importer test to separate file
