+++ tests/src/test/scala/gapt/formats/tip/TipSmtParserTest.scala
@@ -43,20 +43,6 @@ import org.specs2.mutable._
 
 class TipSmtParserTest extends Specification {
 
-  "bin_distrib.smt2" in {
-    val problem =
-      TipSmtImporter.parse( ClasspathInputFile( "bin_distrib.smt2" ) )
-    val one = Const( "One", TBase( "Bin" ) )
-    val oneAnd = Const( "OneAnd", TBase( "Bin" ) ->: TBase( "Bin" ) )
-    val instanceSequent =
-      problem
-        .toSequent
-        .map(
-          identity,
-          instantiate( _, Seq( one, one, oneAnd( oneAnd( one ) ) ) ) )
-    Escargot getResolutionProof instanceSequent must beSome
-  }
-
   "parsing constant declaration" in {
 
     "parsing well-formed constant declaration should succeed" in {
